Move UserFactory into resources.py
Put it in resources.py where all the other factories are. This is just a
cut-paste.
